requirements.txt
robyn>=0.9
uvicorn[standard]>=0.23
passlib[bcrypt]>=1.7
bcrypt<4.0
argon2-cffi>=23.1
aiosqlite>=0.20
markupsafe>=2.1
Jinja2>=3.1
Pillow>=10.0
insightface>=0.7.3
onnxruntime>=1.17
easyocr>=1.7.2
pillow-heif>=0.18
